+++ scripts/generate_setup.py
@@ -17,6 +17,7 @@
 import logging
 import os
 import subprocess
+import sys
 
 from jinja2 import Template
 
@@ -72,7 +73,17 @@ def main():
         with open(generated_file, "w") as fh:
             fh.write(generated)
             fh.flush()
-        subprocess.run(["black", "-q", generated_file], check=True)
+
+    subprocess.run(
+        [
+            sys.executable,
+            "scripts/eachdist.py",
+            "format",
+            "--path",
+            "instrumentation",
+        ],
+        check=True,
+    )
 
 
 if __name__ == "__main__":