+++ scripts/prepare_release.sh
@@ -20,6 +20,8 @@ git checkout -b release/${VERSION}
 git push origin release/${VERSION}
 
 ./scripts/eachdist.py update_versions --versions stable,prerelease
+./scripts/generate_setup.py
+./scripts/generate_instrumentation_bootstrap.py
 rc=$?
 if [ $rc != 0 ]; then
     echo "::set-output name=version_updated::0"